requirements.txt
email-validator>=2.2.0
flask-dance>=7.1.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
pillow>=11.3.0
mysqlclient>=2.2.0
werkzeug>=3.1.3
flask-login>=0.6.3
oauthlib>=3.3.1
pyjwt>=2.10.1
python-dotenv>=1.0.0